+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 -   Mass export multiple sessions or cards as a zip archive.
 -   Mass import multiple sessions or cards from a zip archive.
 -   Emulation menu can now be accessed by right clicking.
+-   Quick switch between session by clicking on the breadcrumb
+    (title) on the emulator tab.
 -   Fix a few instances of dialogs being hidden by a loader.
 -   Other UI fixes performance and optimizations